requirements.txt
streamlit
pandas
pdfplumber
openai
python-dotenv
